+++ pkg/network/tunnel/client/session.go
@@ -22,7 +22,7 @@ import (
 	"emperror.dev/errors"
 	"github.com/xtaci/smux"
 
-	"github.com/cisco-open/nasp/pkg/tunnel/api"
+	"github.com/cisco-open/nasp/pkg/network/tunnel/api"
 )
 
 type session struct {
