+++ pkg/network/tunnel/server/session.go
@@ -26,7 +26,7 @@ import (
 	"github.com/xtaci/smux"
 
 	"github.com/cisco-open/nasp/pkg/network/proxy"
-	"github.com/cisco-open/nasp/pkg/tunnel/api"
+	"github.com/cisco-open/nasp/pkg/network/tunnel/api"
 )
 
 type session struct {
